Cleanup & reorganize files
