--- src/data/data_sources.md
@@ -1,4 +1,0 @@
-
-## Traffic noise zones in Helsinki 2017
-* Downloaded via [Helsinki Region Infoshare](https://hri.fi/data/en_GB/dataset/helsingin-kaupungin-meluselvitys-2017)
-* CC BY 4.0
