--- src/data/tests/test_OD_lines.geojson
@@ -1,22 +1,0 @@
-{
-"type": "FeatureCollection",
-"name": "test_OD_lines",
-"crs": { "type": "name", "properties": { "name": "urn:ogc:def:crs:OGC:1.3:CRS84" } },
-"features": [
-{ "type": "Feature", "properties": { "id": null, "OD": "1" }, "geometry": { "type": "MultiLineString", "coordinates": [ [ [ 24.967648140218643, 60.218764780264301 ], [ 24.970867782581262, 60.223451400385244 ] ] ] } },
-{ "type": "Feature", "properties": { "id": null, "OD": "10" }, "geometry": { "type": "MultiLineString", "coordinates": [ [ [ 24.967021391011894, 60.213418322406227 ], [ 24.959807883643862, 60.220397976144632 ] ] ] } },
-{ "type": "Feature", "properties": { "id": null, "OD": "3" }, "geometry": { "type": "MultiLineString", "coordinates": [ [ [ 24.958331169093253, 60.214063680551078 ], [ 24.968005943764091, 60.207964496608511 ] ] ] } },
-{ "type": "Feature", "properties": { "id": null, "OD": "4" }, "geometry": { "type": "MultiLineString", "coordinates": [ [ [ 24.975648228689565, 60.212726642632852 ], [ 24.96141702258263, 60.214283409157915 ] ] ] } },
-{ "type": "Feature", "properties": { "id": null, "OD": "5" }, "geometry": { "type": "MultiLineString", "coordinates": [ [ [ 24.974972840753988, 60.207473620794126 ], [ 24.953807799941359, 60.211322024290261 ] ] ] } },
-{ "type": "Feature", "properties": { "id": null, "OD": "6" }, "geometry": { "type": "MultiLineString", "coordinates": [ [ [ 24.968856390143337, 60.217338856311031 ], [ 24.968652780124753, 60.210567694708772 ] ] ] } },
-{ "type": "Feature", "properties": { "id": null, "OD": "7" }, "geometry": { "type": "MultiLineString", "coordinates": [ [ [ 24.972537609821941, 60.21524731567137 ], [ 24.958948991572363, 60.215807217877121 ] ] ] } },
-{ "type": "Feature", "properties": { "id": null, "OD": "8" }, "geometry": { "type": "MultiLineString", "coordinates": [ [ [ 24.965273692843969, 60.218705063825773 ], [ 24.958286913527104, 60.214717907014638 ] ] ] } },
-{ "type": "Feature", "properties": { "id": null, "OD": "9" }, "geometry": { "type": "MultiLineString", "coordinates": [ [ [ 24.960227571616286, 60.206898152592984 ], [ 24.967515851952285, 60.21067978000103 ] ] ] } },
-{ "type": "Feature", "properties": { "id": null, "OD": "2" }, "geometry": { "type": "MultiLineString", "coordinates": [ [ [ 24.973109869083839, 60.21575810878474 ], [ 24.972113741491832, 60.214621413321971 ] ] ] } },
-{ "type": "Feature", "properties": { "id": null, "OD": "11" }, "geometry": { "type": "MultiLineString", "coordinates": [ [ [ 24.97240453022598, 60.213964623771652 ], [ 24.972642181967824, 60.213608285549817 ] ] ] } },
-{ "type": "Feature", "properties": { "id": null, "OD": "12" }, "geometry": { "type": "MultiLineString", "coordinates": [ [ [ 24.973849575978527, 60.213390088663296 ], [ 24.974285775919881, 60.213639493241445 ] ] ] } },
-{ "type": "Feature", "properties": { "id": null, "OD": "13" }, "geometry": { "type": "MultiLineString", "coordinates": [ [ [ 24.974739728033896, 60.211910272139789 ], [ 24.976318588237014, 60.212196923598398 ] ] ] } },
-{ "type": "Feature", "properties": { "id": null, "OD": "14" }, "geometry": { "type": "MultiLineString", "coordinates": [ [ [ 24.973123808014076, 60.211787574553988 ], [ 24.97375521188961, 60.211658183442232 ] ] ] } },
-{ "type": "Feature", "properties": { "id": null, "OD": "15" }, "geometry": { "type": "MultiLineString", "coordinates": [ [ [ 24.969663593428624, 60.212528382709991 ], [ 24.965269166096316, 60.208829768571377 ] ] ] } }
-]
-}
