--- src/data/tests/test_walk_line.cpg
@@ -1,1 +1,0 @@
-ISO-8859-1
