--- src/deploy/aqi-processor-app.service
@@ -1,13 +1,0 @@
-[Unit]
-Description=python application for processing Enfuser AQI data
-After=network.target
-
-[Service]
-User=ubuntu
-Group=www-data
-WorkingDirectory=/home/ubuntu/gp-server
-Environment="PATH=/home/ubuntu/miniconda3/envs/aqi-processing/bin"
-ExecStart=/home/ubuntu/miniconda3/envs/aqi-processing/bin/python aqi_processor_app.py
-
-[Install]
-WantedBy=multi-user.target
