--- src/deploy/gp-app.service
@@ -1,13 +1,0 @@
-[Unit]
-Description=Gunicorn instance to serve green path app
-After=network.target
-
-[Service]
-User=ubuntu
-Group=www-data
-WorkingDirectory=/home/ubuntu/gp-server
-Environment="PATH=/home/ubuntu/miniconda3/envs/gp-server-3/bin"
-ExecStart=/home/ubuntu/miniconda3/envs/gp-server-3/bin/gunicorn --workers 8 --timeout 200 --log-level=info --bind unix:green-paths.sock -m 007 green_paths_wsgi:app
-
-[Install]
-WantedBy=multi-user.target
