--- src/env_aqi_processing.yml
@@ -1,16 +1,0 @@
-name: aqi-processing-2
-
-channels:
-  - conda-forge
-dependencies:
-  - python=3.6
-  - pylint
-  - pytest
-  - apscheduler
-  - geopandas
-  - python-igraph
-  - networkx
-  - rasterio
-  - rioxarray
-  - xarray
-  - boto3
