--- src/env_gp_server.yml
@@ -1,16 +1,0 @@
-name: gp-server-3
-
-channels:
-  - conda-forge
-dependencies:
-  - python=3.6
-  - pylint
-  - pytest
-  - apscheduler
-  - geopandas
-  - networkx
-  - python-igraph
-  - flask
-  - flask-cors
-  - flask-testing
-  - gunicorn
