--- src/env_graph_construction.yml
@@ -1,10 +1,0 @@
-name: graph-construction
-
-channels:
-  - conda-forge
-dependencies:
-  - python=3.6
-  - jupyterlab
-  - pylint
-  - geopandas
-  - osmnx
